Make connectivity check more robust
Some endpoints chosen for connectivity check may return an error,
for example, due to rate limits being exceeded. However, as long
as the given website could be reached, it means the connection's
working and fallback to NAT Gateway should not be initiated. Such
occurrences can be simply logged but treated as check success.

Add a `User-Agent` header to connection check requests, which also
prevents some websites from returning an error in the first place.