Treat all http codes < 500 as success (#147)
Generally speaking, as soon as a http code is returned, we successfully
interacted with a server, thus internet connection is available. This is
even true for most non-success status codes, such as 404.

There might be some edge cases, where the final server is not available
and an intermediate server/gateway/proxy/... answers. This probably could
still be within AWS infrastructure, thus we should treat these as failure,
just in case.

These edge cases would return the following http codes:

502: Proxy/gateway received invalid response from upstream
503: Service (or upstream) unavailable; proxy can't forward the request
504: Gateway/proxy timed out waiting for upstream response

Simplified, we fail on any http code >= 500.