Add option to install and configure CloudWatch Agent (#155)
* Add option to install and configure CloudWatch Agent

The CloudWatch Agent can be used to bring visibility into low-level
networking metrics, such as the volume of dropped packets.
The configuration file aims to report most of the metrics provided for managed NAT Gateways
and is adapted from the fck-nat documentation (https://fck-nat.dev/develop/features/#metrics).

* Add credit to the fck-nat project in the user-data script

The cwagent.json.tftpl user-data script generates a configuration file
for the CloudWatch Agent that was published by the fck-nat.dev project
and is adapted from their documentation.

* Enable CloudWatch Agent in Example deployment

This should allow testing this feature.