fix: route table id is being updated with nat gateway id
While updating the logging to switch to structured logging, feedback was
provided that identified a case of inaccurate and broken logging when
`replace_route()` is invoked. Not only is the log message wrong by using
a route table id instead of the route id, but also the use of the
`new_route_table` variable would produce a dictionary in the log.

The log line has now been updated to accurately state what is going on,
and remove the use of a nested dictionary in the structured log.

BREAKING CHANGE: Users monitoring this message with the original
specific phrasing will experience broken monitors. It is recommended
that users should be monitoring generally for ERROR level logs,
or the "Failed connectivity tests! Replacing route" logs specifically.